로그인->메인화면 연결
로그인->메인화면 연결을 하였습니다.
근데 월간 매출 버튼을 누르면 자꾸 오류가 생깁니다.
