+++ .gitignore
@@ -20,3 +20,7 @@
 
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
+/bin/
+.idea/workspace.xml
+.idea/workspace.xml
+.project
