--- src/view/TestFrm.java
@@ -1,12 +1,0 @@
-package view;
-
-import javax.swing.JFrame;
-
-public class TestFrm extends JFrame
-{
-	public TestFrm()
-	{
-		setSize(800, 600);
-		setVisible(true);
-	}
-}
